feat: add mock data for categories and products with associated images
